OK, story time: So, Qt really doesn't like it if you try to send data on a socket in one thread and read data from it in another thread, even if you do your locking correctly. Touching the socket in a thread other than the one in which it was instantiated == death. I started working on a patch that does away with threads and blocking I/O and simply uses Qt's signals and slots for dealing with socket operations asynchronously from the main thread. This patch got really big (already 3000+ lines and I'm not yet done) and isn't something we want to do when we're trying to put out a 'stable' Vidalia bundle. We'll save that patch for the next version (or start playing the stable/dev branch game?).
So, here's a hack much like what was removed in r2580, except this actually
works with Qt 4.4. The basic idea is to post an event to the control socket,
which will get delivered to the socket in the correct thread. The event tells
the socket what to send, and the main thread waits for the result of the send.
Sucky, huh?


svn:r2838