Add the possibility to consult the ControlReply after a SETCONF
This will be used to identify when an error occurs with SETCONF, whether it
is a 553 error or any other kind. If it's a 553, then the option is not
wrong but tor can't apply that change while it is running. Otherwise, it's
an error in the option and it should be treated as such.