A fix for trac bug 6724, where vidalia won't complain about "bridge" anymore.
Now Vidalia will simply strip "bridge" from the front, when adding new bridges
to the list. Also the input was reworked to a multi line text box so one can
add more than one bridge at a time. This is helpful as one could paste all the
bridges from bridges.tpo in one pass. Also there was a small bug in applying
the settings as deleted bridges where not commited to the configuration.